--- apps/web/src/entities/post/index.ts
@@ -1,1 +1,0 @@
-export { getPostList } from "./api";
