--- apps/web/src/widgets/bottomSheet/index.ts
@@ -1,1 +1,0 @@
-export { BottomSheet } from "./ui/BottomSheet";
