--- apps/web/src/widgets/footerNav/index.ts
@@ -1,1 +1,0 @@
-export { FooterNav } from "./ui";
