--- apps/web/src/widgets/header/index.ts
@@ -1,2 +1,0 @@
-export { HomeHeader } from "./ui/HomeHeader";
-export { PostHeader } from "./ui/PostHeader";
