--- apps/web/src/widgets/post/index.ts
@@ -1,1 +1,0 @@
-export { PostList } from "./ui/PostList";
