--- apps/web/src/widgets/tabs/index.ts
@@ -1,1 +1,0 @@
-export { Tabs } from "./ui/Tabs";
